requirements.txt
streamlit
google-generativeai
pandas
chromadb
